refactor(7105): separate verifier policy from report coordination
Extract claim validation into a bounded module, improve front-door diagnostics, and retire the historical rehearsal marker binary after canonical behavior coverage remains green.

Constraint: Touched Rust files and functions must remain within repository size limits

Rejected: Keep the 214-line verifier coordinator | policy ownership was obscured and violated the file budget

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Runtime verifier contracts are authoritative; do not reintroduce a historical evidence-marker test as behavior proof

Tested: focused 23-test suite; strict workspace clippy; touched-size policy; manifest, stale-reference, and command-surface checks

Not-tested: Full make test stopped only at the pre-commit governance-history ratio guard; rerun required at this committed head
